@@ -1,6 +1,6 @@
 {
   "name": "vue-agile",
-  "version": "1.1.3",
+  "version": "2.0.0",
   "description": "A carousel component for Vue.js",
   "author": "Łukasz Florczak <lukasz@florczak.dev>",
   "scripts": {
@@ -23,22 +23,32 @@
     "lodash.throttle": "^4.1.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.6.0",
-    "@vue/cli-plugin-eslint": "^3.6.0",
-    "@vue/cli-plugin-unit-jest": "^4.3.1",
-    "@vue/cli-service": "^4.3.1",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/test-utils": "1.0.0-beta.31",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.16.0",
-    "eslint-plugin-pug": "^1.2.2",
-    "eslint-plugin-vue": "^5.0.0",
-    "pug": "^2.0.3",
-    "pug-plain-loader": "^1.0.0",
+    "core-js": "^3.6.5",
+    "vue": "^3.0.0",
+    "@types/jest": "^24.0.19",
+    "@typescript-eslint/eslint-plugin": "^4.18.0",
+    "@typescript-eslint/parser": "^4.18.0",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-typescript": "~4.5.0",
+    "@vue/cli-plugin-unit-jest": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "@vue/eslint-config-standard": "^5.1.2",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "@vue/test-utils": "^2.0.0-0",
+    "eslint": "^6.7.2",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-vue": "^7.0.0",
+    "lint-staged": "^9.5.0",
     "sass": "^1.18.0",
     "sass-loader": "^8.0.2",
-    "vue": "^2.6.10",
-    "vue-template-compiler": "^2.5.21"
+    "typescript": "~4.1.5",
+    "vue-cli-plugin-pug": "~2.0.0",
+    "vue-jest": "^5.0.0-0"
   },
   "bugs": {
     "url": "https://github.com/lukaszflorczak/vue-agile/issues"
