@@ -2,26 +2,26 @@
  * Helpers methods
  */
 const mixin = {
-	methods: {
-		/**
-		 * Set window & container width
-		 */
-		getWidth () {
-			if (this.isSSR) {
-				return false
-			}
+  methods: {
+    /**
+     * Set window & container width
+     */
+    getWidth () {
+      if (this.isSSR) {
+        return false
+      }
 
-			this.widthWindow = window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth
-			this.widthContainer = this.$refs.list.clientWidth
-		},
+      this.widthWindow = window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth
+      this.widthContainer = this.$refs.list.clientWidth
+    },
 
-		/**
-		 * Convert HTML Collection to JS Array
-		 */
-		htmlCollectionToArray (collection) {
-			return Array.prototype.slice.call(collection, 0)
-		}
-	}
+    /**
+     * Convert HTML Collection to JS Array
+     */
+    htmlCollectionToArray (collection) {
+      return Array.prototype.slice.call(collection, 0)
+    }
+  }
 }
 
 export default mixin
